OpenAI_API/requirements.txt
Flask
flask-cors
openai
yfinance
phidata
python-dotenv
flask-caching
